requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
Django==4.2.16
django-allauth==65.0.2
django-browser-reload==1.17.0
django-extensions==3.2.3
django-mongoengine==0.5.6
django_csp==3.8
dnspython==2.6.1
google==3.0.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
MarkupSafe==2.1.5
mongoengine==0.29.1
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pymongo==4.10.0
pyparsing==3.1.4
python-dotenv==1.0.1
razorpay==1.4.2
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
watchman==0.0.1
Werkzeug==3.0.4
